Check for S3 image before upload. Delete temp files after upload. Eager load songs

+++ app/models/api_client.rb
@@ -75,12 +75,17 @@ module APIClient
   end
 
   def self.upload_image_to_s3(image_link, spotify_id)
-    file = "/tmp/#{spotify_id}.png"
+    s3_image_url = "https://s3.amazonaws.com/sportcasts.com/#{spotify_id}.png"
+    file = Rails.root.join('tmp', "#{spotify_id}.png").to_s
+    return s3_image_url if S3_BUCKET.object(File.basename(file)).exists?
+
     File.open(file, "wb") do |f|
       f.write HTTParty.get(image_link).body
     end
 
+
     S3_BUCKET.object(File.basename(file)).upload_file(file, acl: 'public-read')
-    "https://s3.amazonaws.com/sportcasts.com/#{spotify_id}.png"
+    File.delete(file) if File.exist?(file)
+    s3_image_url
   end
 end
